+++ packages/ai/src/test/mock-provider-v3.ts
@@ -4,8 +4,8 @@ import {
   LanguageModelV3,
   NoSuchModelError,
   ProviderV3,
-  SpeechModelV2,
-  TranscriptionModelV2,
+  SpeechModelV3,
+  TranscriptionModelV3,
 } from '@ai-sdk/provider';
 
 export class MockProviderV3 implements ProviderV3 {
@@ -25,8 +25,8 @@ export class MockProviderV3 implements ProviderV3 {
     languageModels?: Record<string, LanguageModelV3>;
     embeddingModels?: Record<string, EmbeddingModelV3<string>>;
     imageModels?: Record<string, ImageModelV3>;
-    transcriptionModels?: Record<string, TranscriptionModelV2>;
-    speechModels?: Record<string, SpeechModelV2>;
+    transcriptionModels?: Record<string, TranscriptionModelV3>;
+    speechModels?: Record<string, SpeechModelV3>;
   } = {}) {
     this.languageModel = (modelId: string) => {
       if (!languageModels?.[modelId]) {
@@ -58,7 +58,7 @@ export class MockProviderV3 implements ProviderV3 {
       }
       return transcriptionModels[modelId];
     };
-    this.speechModel = (modelId: string) => {
+    this.speechModel = (modelId: string): SpeechModelV3 => {
       if (!speechModels?.[modelId]) {
         throw new NoSuchModelError({ modelId, modelType: 'speechModel' });
       }